Refactor Hand as an abstract class with two game hand classes inheriting varying ViewGroup methods. Beginning of solitaire animations (xml layouts).
