+++ CardGameSuiteApp/app/src/main/java/com/example/cardgamesuiteapp/display/Hand.java
@@ -10,7 +10,7 @@ import android.view.WindowManager;
 
 import java.util.ArrayList;
 
-public class Hand extends ViewGroup {
+abstract public class Hand extends ViewGroup {
     int deviceWidth;
     ArrayList<Card> cards = new ArrayList<Card>();
 
@@ -27,7 +27,7 @@ public class Hand extends ViewGroup {
         init(context);
     }
 
-    private void init(Context context) {
+    void init(Context context) {
         final Display display = ((WindowManager) context.getSystemService(Context.WINDOW_SERVICE)).getDefaultDisplay();
         Point deviceDisplay = new Point();
         display.getSize(deviceDisplay);
@@ -62,25 +62,7 @@ public class Hand extends ViewGroup {
     }
 
     @Override
-    protected void onLayout(boolean changed, int left, int top, int right, int bottom) {
-        final int count = getChildCount();
-
-        int cardWidth = getMeasuredWidth() / 2;
-        int cardHeight = getMeasuredHeight() / 2;
-
-        for (int i = 0; i < count; i++) {
-            View child = getChildAt(i);
-
-            if (child.getVisibility() == GONE) {
-                return;
-            }
-            left = cardWidth * (i % 2);
-            top = cardHeight * (i / 2);
-            right = cardWidth * (i % 2 + 1);
-            bottom = cardHeight * (i / 2 + 1);
-            child.layout(left, top, right, bottom);
-        }
-    }
+    abstract protected void onLayout(boolean changed, int left, int top, int right, int bottom);
 
     @Override
     protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
