+++ CardGameSuiteApp/app/src/main/res/layout/fives_six_players.xml
@@ -57,7 +57,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent" />
 
-    <com.example.cardgamesuiteapp.display.Hand
+    <com.example.cardgamesuiteapp.display.FivesHand
         android:id="@+id/player1"
         android:layout_width="150dp"
         android:layout_height="210dp"
@@ -67,7 +67,7 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.99" />
 
-    <com.example.cardgamesuiteapp.display.Hand
+    <com.example.cardgamesuiteapp.display.FivesHand
         android:id="@+id/player2"
         android:layout_width="110dp"
         android:layout_height="154dp"
@@ -78,7 +78,7 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.654" />
 
-    <com.example.cardgamesuiteapp.display.Hand
+    <com.example.cardgamesuiteapp.display.FivesHand
         android:id="@+id/player3"
         android:layout_width="110dp"
         android:layout_height="154dp"
@@ -89,7 +89,7 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.027" />
 
-    <com.example.cardgamesuiteapp.display.Hand
+    <com.example.cardgamesuiteapp.display.FivesHand
         android:id="@+id/player4"
         android:layout_width="110dp"
         android:layout_height="154dp"
@@ -100,7 +100,7 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.027" />
 
-    <com.example.cardgamesuiteapp.display.Hand
+    <com.example.cardgamesuiteapp.display.FivesHand
         android:id="@+id/player5"
         android:layout_width="110dp"
         android:layout_height="154dp"
@@ -111,7 +111,7 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.027" />
 
-    <com.example.cardgamesuiteapp.display.Hand
+    <com.example.cardgamesuiteapp.display.FivesHand
         android:id="@+id/player6"
         android:layout_width="110dp"
         android:layout_height="154dp"
